+++ exportplan/forms.py
@@ -280,8 +280,8 @@ class ExportPlanTargetMarketsResearchForm(forms.Form):
 
 
 class ExportPlanBusinessObjectivesForm(forms.Form):
-    rational = forms.CharField(
-        label='Business performance',
+    rationale = forms.CharField(
+        label='Add business rationale',
         required=False,
         widget=Textarea(attrs={'placeholder': 'Add some text'})
     )